requirements.txt
annotated-types==0.7.0
anyio==3.7.1
botocore==1.35.44
cachetools==5.5.0
cement==2.10.14
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
dnspython==2.6.1
easyocr==1.7.2
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.16.1
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.4
httpx==0.27.0
idna==3.7
imageio==2.36.1
Jinja2==3.1.4
jmespath==1.0.1
langdetect==1.0.9
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.1
ocrspace==2.3.0
openai==1.37.0
opencv-python-headless==4.11.0.86
packaging==24.1
pathspec==0.10.1
pillow==10.4.0
pyclipper==1.3.0.post6
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pytesseract==0.3.13
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
scikit-image==0.25.0
scipy==1.15.1
semantic-version==2.10.0
setuptools==75.2.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.13.1
termcolor==2.5.0
tifffile==2025.1.10
tiktoken==0.8.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.30.3
watchfiles==0.24.0
watchgod==0.8.2
wcwidth==0.2.13
websockets==12.0